Updates zxc to v0.2.0
Updates the zxc compression library to version 0.2.0.

This commit integrates the new version of the zxc library,
including necessary header files, source files, and updates
the compressor description in `lzbench.h`. The update also
reflects the version change in the CHANGELOG.

This change replaces the previous zxc implementation to align
with the updated API and features of version 0.2.0.
